added descriptions of tables & columns with unique, null and freshness tests

+++ models/staging/stg_raw__product.sql
@@ -10,7 +10,7 @@
 
      select
          products_id,
-         CAST(purchSE_PRICE AS FLOAT) AS purchase_price
+         CAST(purchSE_PRICE AS FLOAT64) AS purchase_price
 
 
      from source
